+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Indestructible trait, prevents durability loss (not found on any default part for now, but the trait does work)
+### Changed
+- Broken armor no longer renders on the player
 
 ## [1.6.3] - 2020-05-07
 ### Changed